DRY findLSB
